% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{methods}
\name{Arith,Polynomial,Polynomial-method}
\alias{Arith,Polynomial,Polynomial-method}
\title{Operations Arithmetiques entre Polynomes ou Polynome et Nombre}
\usage{
\S4method{Arith}{Polynomial,Polynomial}(e1, e2)
}
\description{
Cette méthode S4 rendant possible les opérations arithmétiques entre polynômes ou polynôme et nombre, est une méthode
du groupe générique \code{\link[methods]{Arith}}.
}
\details{
Les opérations suivantes sont implémentées :

\code{`+`}, \code{sum()} : addition polynomiale ou opératon unaire

\code{`-`} : soustraction polynomiale ou calcul de l'opposé d'un polynôme

\code{`*`} : multiplication de deux polynômes, un polynôme par un nombre ou encore un nombre par un polynôme

\code{`\%\%`} et \code{`\%\\\%`} : division euclidienne entre polynômes (diviseur non nul), le premier étant le quotient et le dernier le résidu

\code{`/`} : division d'un polynôme par un nombre non nul.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
poly0 <- Polynomial(c(0,0,0))

## Addition
poly1 + poly2
+ poly2
+ poly0
sum(list(poly1, poly1, poly1, poly1))

## Soustraction
poly1 - poly2
- poly2
- poly0
poly0 - poly0
poly0 + poly0
poly0 * poly1

## Multipliacation
poly1 * poly2
poly0 * poly0
poly0 * poly1
poly2 * 2
2 * poly2

## Calcul du quotient
Polynomial(c(1, -1, 1), roots = TRUE) \%\% Polynomial(c(1, 1), roots = TRUE)
Polynomial(c(1, 0, -2, 0, 0, 3, -5, -6, -8)) \%\% Polynomial(c(3, -5, -6, -8))
## Cette commande renvoyera une erreur
#poly1 \%\% poly0

## Calcul du residu
Polynomial(c(1, 3, -1, 1), roots = TRUE) \%/\% Polynomial(c(5, 10, 5))
Polynomial(c(1, 0, -2, 0, 0, 3, -5, -6, -8)) \%/\% Polynomial(c(3, -5, -6, -8))
## Cette commande renvoyera une erreur
#poly1 \%\% poly0

## Division par un nombre
poly2 / 2
}
\seealso{
Voir \code{\link[methods]{Arith}} pour plus d'informations.
}
