% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{methods}
\name{as.character,Polynomial-method}
\alias{as.character,Polynomial-method}
\title{S4 Methode pour convertir un polynome en une chaine de caracteres ou en une expression}
\usage{
as.character(x, "Polynomial")
}
\description{
Méthode S4 de \code{as.character.Polynomial}. Seule différence c'est la forme d'appel.
Voir \code{\link{as.character.Polynomial}}.
}
