% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{as.character.Polynomial}
\alias{as.character.Polynomial}
\title{Convertir un polynome en une chaine de caracteres ou en une expression.}
\usage{
\method{as.character}{Polynomial}(x, expressionStyle = FALSE)
}
\arguments{
\item{x}{un objet \code{Polynomial}.}

\item{expressionStyle}{Si on veut un format humain (\code{FALSE}, par défaut) ou d'expression (\code{TRUE}).}
}
\value{
Selon le choix, une chaîne de caractères ou une expression.
}
\description{
Convertir un objet de classe \code{Polynomial} soit en une chaîne de caractères ou en une expression, selon besoin.
}
\details{
Quand \code{expressionStyle} est \code{FALSE}, la fonction renvoie un polynôme au format d'après l'habitude de l'humain,
c'est-à-dire sans signe de multiplication. Par contre quand  \code{expressionStyle} est \code{TRUE}, la fonction va retourner
une expression au format TEX comme si l'on utilisait la fonction \code{expression()}. La différence c'est que cette fois-ci
il y a des signes de multiplications entre les coeffs et les termes.
}
\examples{
as.character.Polynomial(Polynomial(c(0, 1)))
as.character.Polynomial(Polynomial(c(1, -1)))
as.character.Polynomial(Polynomial(c(0, 1, -1, 1, 0, -1, 2, 0)), expressionStyle = TRUE)

}
