% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{basePolynomial}
\alias{basePolynomial}
\title{Creation d'une Base Polynomiale Ortho-normee ou Non}
\usage{
basePolynomial(d, ortho = FALSE)
}
\arguments{
\item{d}{Nombre entier positif, précisant la dimension de la base.}

\item{ortho}{Logique, \code{FALSE} par défaut. Si on veut que la base polnomiale obtenue soit ortho-normée.}
}
\value{
Une liste dont chaque membre contient une composante de la base, qui est un objet \code{Polynomial}.
}
\description{
Créer une base polynomiale sur [-1, 1] d'une dimension quelconque. Ortho-normée ou non.
}
\details{
Si \code{ortho} est \code{FALSE}, \code{1}, \code{x}, \code{x^2}... seront renoyés. Si \code{ortho} est \code{TRUE},
la fonction renvoie une base ortho-normée de dimension \code{d} en utilisant l'algorithme de Gram-Schmidt.
}
\examples{
basePolynomial(3)
BasePoly4 <- basePolynomial(4, ortho=TRUE)
## Est-cette base vraiment orthonormée ?
BasePoly4
sapply(BasePoly4, function(x) integratePolynomial(x*x, -1, 1)) # doit renvoyer : [1] 1 1 1 1
apply(combn(BasePoly4, 2), 2, function(x) integratePolynomial(x[[1]]*x[[2]], -1, 1)) # doit renvoyer : [1] 0 0 0 0 0 0

}
