% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{degre}
\alias{degre}
\title{Calcul du Degre d'un Polynome}
\usage{
degre(p)
}
\arguments{
\item{p}{Un objet de classe \code{Polynomial}.}
}
\value{
Un entier positif. Pour le polynôme nul, elle renvoie -Inf.
}
\description{
Calcul du Degré d'un Polynôme.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
degre(poly1)
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
degre(poly2)
poly0 <- Polynomial(c(0, 0, 0))
degre(poly0)
}
