% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{derivePolynomial}
\alias{derivePolynomial}
\title{Calcul de la Derivee d'un Polynome}
\usage{
derivePolynomial(x)
}
\arguments{
\item{x}{Un objet de classe \code{Polynomial}.}
}
\value{
Un objet de classe \code{Polynomial}.
}
\description{
Calcul de la dérivée d'un polynôme.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
derivePolynomial(poly1)
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
derivePolynomial(poly2)
poly0 <- Polynomial(c(0,0,0))
derivePolynomial(poly0)

}
