% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{methods}
\name{is,Polynomial-method}
\alias{is,Polynomial-method}
\title{Verification de Classe \code{Polynomial}}
\usage{
\S4method{is}{Polynomial}(object)
}
\arguments{
\item{x}{Objet à vérifier si c'est de classe \code{Polynomial}.}
}
\value{
la fonction \code{is.Polynomial} ou \code{is(x, "Polynomial")} renvoie un booléen.
}
\description{
Vérifier si un objet est de classe \code{Polynomial}.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
is.Polynomial(poly1)
is(poly1, "Polynomial")
is.Polynomial(c(1, 2))
}
