% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{methods}
\name{plot,Polynomial-method}
\alias{plot,Polynomial-method}
\title{Tracer un Polynome}
\usage{
\S4method{plot}{Polynomial}(x, xlim = c(-1, 1),
  main = "Trace polyn<U+00F4>miale", xlab = parse(text =
  as.character.Polynomial(x, expressionStyle = TRUE)), ylab = "valeur",
  col = "darkblue", lwd = 2, ...)
}
\arguments{
\item{x}{Un objet de classe \code{Polynomial}.}

\item{xlim}{L'amplitude de l'axe de X.}

\item{main}{Titre principal.}

\item{xlab}{sous-titre pour l'axe de X. Par défaut, c'est l'expression du polynôme.}

\item{ylab}{sous-titre pour l'axe de Y.}

\item{col}{couleur du polynôme tracé.}

\item{lwd}{l'épaisseur de la ligne.}

\item{...}{tout autre paramètre pouvant être acceptée par la fonction \code{print.function()}.}
}
\description{
Une méthode S4 de la fonction générique \code{plot} pour tracer un polynôme.
}
\details{
La méthode employée au fond c'est \code{plot.function()} et NON PAS \code{plot.xy()}.
}
\examples{
poly1 <- Polynomial(c(1,-2,3))
plot.Polynomial(poly1)
poly0 <- Polynomial(c(0,0,0))
plot(poly0, col='red', lwd=1.5)
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
plot(poly2, col="darkorange", lwd=3, xlim=c(-10, 10))

}
