% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{class}
\name{Polynomial-class}
\alias{Polynomial-class}
\title{Classe Polynomiale}
\description{
Classe S4.
}
\section{Slots}{

\describe{
\item{\code{coeffs}}{Objet de classe \code{"numeric"}. Il est le vecteur enregistrant les coeffs du polynôme dans l'ordre croissant des termes.}

\item{\code{racines}}{Objet de classe \code{"numeric"}. Il est le vecteur enregistrant les racines du polynôme. Sans ordre. \code{NA} par défaut.}
}}

\note{
Accessible par l'utilisateur.
}
