% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{Polynomial}
\alias{Polynomial}
\title{Creer un Nouveau Polynome}
\usage{
Polynomial(num, roots = FALSE, dominant = 1)
}
\arguments{
\item{num}{un vecteur numérique indiquant soit les coefficents soit les racines du polynôme à créer.}

\item{dominant}{numérique atomique, indiquant le coeff du plus haut terme. Ne fonctionne que quand \code{root == TRUE}.}

\item{root}{logique, vaut \code{TRUE} si \code{nom} doit être les racines et \code{FALSE} si coeffs.}
}
\value{
Un objet de classe \code{Polynomial}. Si le polynôme est construit par ses coeffs, le slot \code{racines} sera de valeur \code{NA}.
}
\description{
Création d'un nouveau polynôme soit par ses coefficients, soit par toutes ses racines.
}
\details{
Fonction pour générer un polynôme. Si \code{root} est \code{TRUE}, le vecteur \code{num} est considéré comme les racines du polynôme à créer.
Au cas contraire, il est les coefficients du polynôme, d'où ceux des terms nuls doivent être representé par un 0 et non pas sautés.
Si \code{root == FALSE} et un \code{dominant} sont donnés en même temps, le dernier sera ignoré.
Dans ce cas-là le polynôme sera construit uniquement par ses coefficients, i.e. le vecteur \code{num}.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
poly1
poly3 <- Polynomial(c(1,2), roots = TRUE)
poly3
poly4 <- Polynomial(c(-1,-1), roots = TRUE, dominant = 5)
poly4
}
