% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{primitivePolynomial}
\alias{primitivePolynomial}
\title{Calcul de la Primitive d'un Polynome}
\usage{
primitivePolynomial(p)
}
\arguments{
\item{x}{Un objet de classe \code{Polynomial}.}
}
\value{
Un objet de classe \code{Polynomial}.
}
\description{
Calcul d'une primitive d'un polynôme.
}
\details{
La primitive retournée c'est celle avec la contante nulle.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
primitivePolynomial(poly1)
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
primitivePolynomial(poly2)
poly0 <- Polynomial(c(0,0,0))
primitivePolynomial(poly0)

}
