% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{methods}
\name{summary,Polynomial-method}
\alias{summary,Polynomial-method}
\title{Resumer un Objet de Classe \code{Polynomial}}
\usage{
\S4method{summary}{Polynomial}(object)
}
\arguments{
\item{object}{Un objet \code{Polynomial}.}
}
\value{
Une liste à deux éléments, chacun comprend un vecteur représentant resp. les coeffs. et les racines.
}
\description{
Une méthode S4 de la fonction générique summary() afin de faire afficher les coefficients et les racines d'un polynôme,
informations qui pourront par la suite être extraites par le signe $.
}
\examples{
poly3 <- Polynomial(c(1,2), roots = TRUE)
summary(poly3)

}
