% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{values}
\alias{values}
\title{Evaluation d'un Polynome}
\usage{
values(p, x)
}
\arguments{
\item{p}{Un objet \code{Polynomial}.}

\item{x}{Un vecteur numérique. Les points en lesquels la fonction doit calculer les valeurs du polynôme.}
}
\value{
Un vecteur numériques correspondant aux valeurs cherchées.
}
\description{
Evaluation d'un polynôme en un ou plusieurs points.
}
\examples{
poly1 <- Polynomial(c(1, -2, 3))
values(p=poly1, x=c(1, 2))
poly2 <- Polynomial(c(1,-3,2,0,-5,0,0,0))
values(poly2, c(0, -1, -2))
polyà <- Polynomial(c(0,0,0))
values(poly0, c(0, 0))

}
